UniApi/requirements.txt
appdirs==1.4.4
apt-xapian-index==0.49
apturl==0.5.2
asgiref==3.5.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.3.2
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.5.0
distro-info==1.0
Django==4.0.3
django-ckeditor==6.4.0
django-cors-headers==3.13.0
django-datetimepicker==3.14
django-embed-video==1.4.4
django-filter==21.1
django-js-asset==2.0.0
djangorestframework==3.13.1
duplicity==0.8.20
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
galternatives==1.0.8
gyp==0.1
html5lib==1.1
httplib2==0.18.1
idna==2.10
importlib-metadata==4.0.1
jeepney==0.7.1
keyring==23.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.18.0
lubuntu-update-notifier==0.4.dev0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.9
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
packaging==20.9
paramiko==2.7.2
pexpect==4.8.0
Pillow==8.1.2
pipenv==11.9.0
protobuf==3.12.4
pycairo==1.16.2
pycups==2.0.1
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.9.0
pyRFC3339==1.1
python-apt==2.2.1
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.25.1
scour==0.38.2
SecretStorage==3.3.1
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
systemd-python==234
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.4.6+ds
virtualenv-clone==0.3.0
wadllib==1.3.5
webencodings==0.5.1
xdg==5
xkit==0.0.0
zipp==1.0.0
